Rename core/envmigration to core/modelmigration
