+++ core/modelmigration/package_test.go
@@ -1,7 +1,7 @@
 // Copyright 2015 Canonical Ltd.
 // Licensed under the AGPLv3, see LICENCE file for details.
 
-package envmigration_test
+package modelmigration_test
 
 import (
 	"testing"
