+++ core/modelmigration/phase_internal_test.go
@@ -1,7 +1,7 @@
 // Copyright 2016 Canonical Ltd.
 // Licensed under the AGPLv3, see LICENCE file for details.
 
-package envmigration
+package modelmigration
 
 import (
 	"github.com/juju/utils/set"
