+++ core/modelmigration/phase_test.go
@@ -1,13 +1,13 @@
 // Copyright 2016 Canonical Ltd.
 // Licensed under the AGPLv3, see LICENCE file for details.
 
-package envmigration_test
+package modelmigration_test
 
 import (
 	jc "github.com/juju/testing/checkers"
 	gc "gopkg.in/check.v1"
 
-	migration "github.com/juju/juju/core/envmigration"
+	migration "github.com/juju/juju/core/modelmigration"
 	coretesting "github.com/juju/juju/testing"
 )
